# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CaReMind API
  version: 1.0.0
  description: REST API for private vehicle, maintenance and operating-cost management.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.car-remind.gr/api
    description: Production
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [Vehicles]
      summary: List the signed-in user's vehicles
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Vehicle list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Vehicle' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createVehicle
      tags: [Vehicles]
      summary: Add a vehicle
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/Vehicle' }
      responses:
        '201': { description: Vehicle created }
        '400': { $ref: '#/components/responses/BadRequest' }
  /vehicles/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      operationId: updateVehicle
      tags: [Vehicles]
      summary: Update an owned vehicle
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/Vehicle' }
      responses:
        '200': { description: Vehicle updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteVehicle
      tags: [Vehicles]
      summary: Delete an owned vehicle and its dependent records
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Vehicle deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found or not owned by this user
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  requestBodies:
...
    Vehicle:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Vehicle' }
...
  schemas:
    Error:
      type: object
      properties: { error: { type: string }, code: { type: string } }
...
    Vehicle:
      type: object
      required: [vehicleType, chassisNumber]
      properties:
        id: { type: integer, readOnly: true }
        vehicleType: { type: string, maxLength: 100 }
        chassisNumber: { type: string, maxLength: 50 }
        model: { type: [string, 'null'], maxLength: 100 }
        year: { type: [integer, 'null'], minimum: 1886 }
        currentMileage: { type: [integer, 'null'], minimum: 0 }
...